+++ test/hexapdf/document/test_signatures.rb
@@ -208,6 +208,13 @@ describe HexaPDF::Document::Signatures do
       assert_equal([0, 968, 3590, 2412], sig[:ByteRange].value)
     end
 
+    it "works if the signature object is the last object of the xref section" do
+      field = @doc.acro_form(create: true).create_signature_field('Signature2')
+      field.create_widget(@doc.pages[0], Rect: [0, 0, 0, 0])
+      sig = @doc.signatures.add(@io, @handler, signature: field, write_options: {update_fields: false})
+      assert_equal([0, 3063, 5685, 308], sig[:ByteRange].value)
+    end
+
     it "allows writing to a file in addition to writing to an IO" do
       tempfile = Tempfile.new('hexapdf-signature')
       tempfile.close